third_party/WebKit/LayoutTests/fast/dom/MutationObserver/document-fragment-insertion-expected.txt

Inserting DocumentFragments should remove all children of the fragment before inserting the children.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing appendChild
PASS mutations.length is 2
PASS mutations[0].addedNodes.length is 0
PASS mutations[0].removedNodes.length is 2
PASS mutations[1].addedNodes.length is 2
PASS mutations[1].removedNodes.length is 0


Testing insertBefore
PASS mutations.length is 2
PASS mutations[0].addedNodes.length is 0
PASS mutations[0].removedNodes.length is 2
PASS mutations[1].addedNodes.length is 2
PASS mutations[1].removedNodes.length is 0


Testing replaceChild
PASS mutations.length is 2
PASS mutations[0].addedNodes.length is 0
PASS mutations[0].removedNodes.length is 2
PASS mutations[1].addedNodes.length is 2
PASS mutations[1].removedNodes.length is 1


PASS successfullyParsed is true

TEST COMPLETE
